pyproject.toml
[project]
name = "py"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "ecdsa (>=0.19.1,<0.20.0)",
    "pycryptodome (>=3.23.0,<4.0.0)",
    "uv (>=0.10.8,<0.11.0)",
    "requests (>=2.32.5,<3.0.0)",
    "ruff (>=0.15.5,<0.16.0)",
    "ipython (>=9.11.0,<10.0.0)",
    "sortedcontainers (>=2.4.0,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "cryptography (>=46.0.5,<47.0.0)",
    "pynacl (>=1.6.2,<2.0.0)",
    "isort (>=8.0.1,<9.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
